+++ src/components/searchNavBar/index.tsx
@@ -1,6 +1,13 @@
 import { Container, Navbar, Form } from 'react-bootstrap';
+import { useState } from 'react';
 
-const SearchNavBar = () => {
+const SearchNavBar = ({
+	onSearchChange, search_term
+}: {
+	onSearchChange: EventTarget | any;
+	search_term: string
+}) => {
+
 	return (
 		<>
 			<Navbar
@@ -16,6 +23,9 @@ const SearchNavBar = () => {
 					<Navbar.Collapse id="home-navbar">
 						<Form className="m-auto w-100">
 							<Form.Control
+								onChange={onSearchChange}
+								name="search_term"
+								value={search_term}
 								placeholder="What would you like to watch?"
 								className="me-2 border-0"
 								aria-label="Search"
